fixing some errors
